fix error with emitting 0 in StreamBuilder and cleaning in main screen
